// libs/core/webhook/src/lib/generated/prisma-client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck
/**
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */
export const WebhookRole = {
  Admin: 'Admin',
  User: 'User',
} as const;

export type WebhookRole = (typeof WebhookRole)[keyof typeof WebhookRole];

export const WebhookStatus = {
  Pending: 'Pending',
  Process: 'Process',
  Success: 'Success',
  Error: 'Error',
  Timeout: 'Timeout',
} as const;

export type WebhookStatus = (typeof WebhookStatus)[keyof typeof WebhookStatus];